total refactoring of css classes naming and folder structure. BEM IS MY FAVORITE METHODOLOGY
