+++ public/blocks/feed-center/template.hbs
@@ -1,0 +1,19 @@
+<div class="feed-center">
+    <img class="feed-center__photos" src="{{ profile.photos }}" alt="{{ profile.photos }}">
+    <!--TODO: image options-footer here-->
+    <div class="feed-center__profile">
+        <div>
+            <h1>{{ profile.name }}{{#if profile.age}},  {{profile.age}}{{/if}}</h1>
+            <div class="feed-center__about">{{ profile.about }}</div>
+        </div>
+        <div class="feed-center__tags">
+            {{#each profile.tags}}
+                <div class="tag tag_size_middle">{{ this.title }}</div>
+            {{/each}}
+        </div>
+    </div>
+    <div class="feed-center__action-buttons">
+        <button class="re_btn__reject">ДА</button>
+        <button class="re_btn__approve">НЕТ</button>
+    </div>
+</div>
