+++ public/blocks/feed-right/feed-right.scss
@@ -1,0 +1,18 @@
+@import "../../modules/variables";
+
+.feed-right {
+  &__title {
+    padding-bottom: $re--small-padding;
+  }
+
+  &__helper {
+    text-align: left;
+    display: flex;
+    justify-content: space-between;
+    align-items: center;
+    margin-bottom: 10px;
+    &_blue {
+      color: $color-blue;
+    }
+  }
+}
