+++ public/blocks/feed-right/template.hbs
@@ -1,0 +1,25 @@
+<div class="feed-right">
+    {{#if isEvents}}
+        {{#if users}}
+            <div class="feed-right__title">
+                <div class="font font_bold font__color_black font__size_small-xx">Уже идут</div>
+            </div>
+            {{#each users as |user-card|}}
+                {{> user-card }}
+            {{/each}}
+        {{else}}
+            <h3 class="feed-right__helper feed-right__helper_blue">"Подписчиков нет"</h3>
+        {{/if}}
+    {{else}}
+        {{#if events}}
+            <div class="feed-right__title">
+                <div class="font font_bold font__color_black font__size_small-xx">Персональные эвенты</div>
+            </div>
+            {{#each events as |big-event|}}
+                {{> big-event }}
+            {{/each}}
+        {{else}}
+            <h3 class="feed-right__helper feed-right__helper_blue">"Эвентов нет :("</h3>
+        {{/if}}
+    {{/if}}
+</div>
