+++ public/blocks/photo-columns/photo-columns.scss
@@ -1,6 +1,6 @@
 @import "../../modules/variables";
 
-.photo_columns {
+.photo-columns {
   width: 100%;
   max-height: 100%;
   column-count: 2;
@@ -10,7 +10,7 @@
   -webkit-column-count: 2;
   -webkit-column-gap: 7px;
 
-  & .photo {
+  &__photo {
     width: 100%;
     height: auto;
     margin-bottom: 3px;
