+++ public/blocks/profile-left/template.hbs
@@ -1,0 +1,60 @@
+<div class="profile-left">
+    <h2>{{ profile.name }}{{#if profile.age}}, {{ profile.age }}{{/if}}</h2>
+    <form action="submit" id="form">
+        <div class="profile-left__scroll">
+            <img class="profile-left__avatar profile-left__avatar_center" src="{{ url }}/{{ profile.avatar.path }}" alt="avatar">
+            <div class="profile-left__options-field">
+                <h3>О себе</h3>
+                <textarea class="profile-left__options-field_textarea" name="about" rows="10"
+                          placeholder="Расскажите о себе и своих увлечениях">{{ profile.about }}</textarea>
+            </div>
+            <div class="profile-left__options-field">
+                <div class="profile-left__header">
+                    <h3>Ваши тэги</h3>
+                    <a class="tags_redirect">Добавить</a>
+                </div>
+                <div class="profile-left__tags">
+                    {{#if profile.tags}}
+                        {{#each profile.tags as |tag|}}
+                            {{>tag}}
+                        {{/each}}
+                    {{else}}
+                        <div class="profile-left__center">
+                            <span class="font font_bold font__size_small font__color_lg">У вас пока нет ни одного тэга</span>
+                        </div>
+                    {{/if}}
+                </div>
+            </div>
+            <div class="profile-left__options-field">
+                <div class="profile-left__header">
+                    <h3>Социальные сети</h3>
+                    <a href="#">Добавить</a>
+                </div>
+                {{#if profile.networks}}
+                    {{#each profile.networks}}
+                        <!-- TODO: add networks -->
+                    {{/each}}
+                {{else}}
+                    <div class="profile-left__center">
+                        <span class="icon icon__size_middle icon_color_lg icon__fb"></span>
+                        <span class="icon icon__size_middle icon_color_lg icon__inst"></span>
+                        <span class="icon icon__size_middle icon_color_lg icon__twitter"></span>
+                        <span class="icon icon__size_middle icon_color_lg icon__vk"></span>
+                        <span class="icon icon__size_middle icon_color_lg icon__ok"></span>
+                        <span class="icon icon__size_middle icon_color_lg icon__moimir"></span>
+                    </div>
+                {{/if}}
+            </div>
+            <div class="profile-left__center">
+                <button class="re_btn re_btn__outline logout">Выйти</button>
+            </div>
+        </div>
+
+        <!-- TODO: add hider for footer shadow -->
+
+        <div class="options-footer">
+            <button class="re_btn re_btn__filled">Сохранить</button>
+            <button class="re_btn re_btn__outline kek">Настройки</button>
+        </div>
+    </form>
+</div>
