+++ public/blocks/tag/tag.scss
@@ -47,7 +47,7 @@
   }
 }
 
-.search_tag {
+.search-tag {
   display: flex;
   justify-content: right;
   align-items: center;
