+++ public/components/my/my.scss
@@ -1,18 +1,16 @@
 @import "../../modules/variables";
 
 .my {
-  &__base {
-    width: 100%;
-    height: 88.5vh; // TODO: bullshit
-    margin: 10px 0;
-    border-radius: $re--border-radius;
-    background-color: $color-white;
-    box-shadow:  20px 20px 60px #cdcdcd,
-             -20px -20px 60px #ffffff;
-    display: flex;
-  }
+  width: 100%;
+  height: 88.5vh; // TODO: bullshit
+  margin: 10px 0;
+  border-radius: $re--border-radius;
+  background-color: $color-white;
+  box-shadow: 20px 20px 60px #cdcdcd,
+  -20px -20px 60px #ffffff;
+  display: flex;
 
-  &__left_column {
+  &__left-column {
     width: 25%;
     max-height: 100%;
     box-sizing: border-box;
@@ -28,26 +26,26 @@
     -webkit-border-bottom-left-radius: $re--border-radius;
     -moz-border-radius-topleft: $re--border-radius;
     -moz-border-radius-bottomleft: $re--border-radius;
-
-    &__header {
-      height: 88px;
-      flex: 0 0 auto;
-      border-bottom: 1px solid $re--color-light-gray;
-      display: flex;
-      justify-content: space-evenly;
-      align-items: center;
-    }
-
-    &__body {
-      padding: $re--medium-padding $re--medium-padding 0;
-      max-height: calc(100% - 88px);
-      //overflow: auto;
-      position: relative;
-      min-height: 692px;
-    }
   }
 
-  &__main_column {
+  &__left-column-header {
+    height: 88px;
+    flex: 0 0 auto;
+    border-bottom: 1px solid $re--color-light-gray;
+    display: flex;
+    justify-content: space-evenly;
+    align-items: center;
+  }
+
+  &__left-column-body {
+    padding: $re--medium-padding $re--medium-padding 0;
+    max-height: calc(100% - 88px);
+    //overflow: auto;
+    position: relative;
+    min-height: 692px;
+  }
+
+  &__main-column {
     width: 75%;
     max-height: 100%;
     border-radius: $re--border-radius;
@@ -55,7 +53,7 @@
     background-color: $color-white;
     padding: $re--big-padding;
 
-    &__body {
+    &-body {
       width: 100%;
       height: calc(100% - 39px);
       display: flex;
