+++ public/components/my/template.hbs
@@ -1,7 +1,7 @@
 <div class="container">
-    <div class="my__base">
-        <div class="my__left_column">
-            <div class="my__left_column__header">
+    <div class="my">
+        <div class="my__left-column">
+            <div class="my__left-column-header">
                 <div class="circle">
                     <svg id="icon" viewBox="0 -26 512 512" xmlns="http://www.w3.org/2000/svg">
                         <path d="m256 100c-5.5195 0-10 4.4805-10 10s4.4805 10 10 10 10-4.4805 10-10-4.4805-10-10-10z"/>
@@ -26,8 +26,11 @@
                     </svg>
                 </div>
             </div>
-            <div class="my__left_column__body"></div>
+            <div class="my__left-column-body"></div>
         </div>
-        <div class="my__main_column"></div>
+        <div class="my__main-column">
+            <h1>{{ title }}</h1>
+            <div class="my__main-column-body"></div>
+        </div>
     </div>
 </div>
