+++ public/components/profile-main/profile-main.scss
@@ -1,6 +1,6 @@
 @import "../../modules/variables";
 
-.new_profile {
+.profile-main {
   &__photos {
     max-height: 100%;
     overflow-y: auto;
@@ -15,14 +15,14 @@
     overflow-y: auto;
     max-width: 20%;
     min-width: 200px;
+  }
 
-    &__group {
-      display: flex;
-      flex-flow: column;
-      justify-content: flex-start;
-      align-items: center;
-      margin-bottom: $re--big-padding;
-    }
+  &__group {
+    display: flex;
+    flex-flow: column;
+    justify-content: flex-start;
+    align-items: center;
+    margin-bottom: $re--big-padding;
   }
 
   &__separator {
