--- public/components/profile/profile.scss
@@ -1,67 +1,0 @@
-@import "../../modules/variables";
-
-.profile {
-
-  text-align: center;
-
-  &__container {
-    height: 90vh;
-    max-width: 900px;
-    margin: 10px auto;
-    background-color: white;
-    border-radius: 13px;
-    padding: 20px;
-  }
-
-  &__title {
-    text-align: left;
-    color: $color-red;
-
-    &_blue {
-      display: flex;
-      justify-content: space-between;
-      align-items: center;
-      color: $color-blue;
-      margin-bottom: 10px;
-    }
-  }
-
-  &__page {
-    display: flex;
-    text-align: left;
-    column-count: 3;
-    -webkit-column-count: 3;
-    -moz-column-count: 3;
-  }
-
-  &_main {
-    display: flex;
-    height: 100%;
-    align-items: flex-start;
-  }
-
-  &__filed {
-    width: 30%;
-    margin: 10px;
-    display: inline-block;
-    overflow: auto;
-  }
-
-  &__photo {
-    width: 100%;
-    margin-bottom: 5px;
-
-    &_img {
-      width: 100%;
-      height: auto;
-      border-radius: 7px;
-    }
-  }
-
-  &__tags {
-    display: flex;
-    flex-direction: row;
-    flex-wrap: wrap;
-    margin-bottom: 20px;
-  }
-}
