--- public/components/profile/template.hbs
@@ -1,66 +1,0 @@
-<div class="profile">
-    <div class="profile__container">
-        <h1 class="profile__title">{{ username }}</h1>
-        <div class="profile__page">
-            <div class="profile__column profile__filed">
-                <h3 class="profile__title profile__title_blue">{{ title1 }}</h3>
-                {{#with textarea}}
-                    <label>
-                        <textarea class="input input__text_small" placeholder="{{ help }}">{{ about }}</textarea>
-                    </label>
-                {{/with}}
-                {{#if edit}}
-                    <div class="profile__btn">
-                        <button class="btn btn_color_ok btn_size_middle" type="submit">
-                            <div class="btn__text btn__text_w">{{ buttonOk }}</div>
-                        </button>
-                        <button class="btn btn_color_b btn_size_middle">
-                            <div class="btn__text btn__text_w">{{ buttonSettings }}</div>
-                        </button>
-                    </div>
-                {{/if}}
-            </div>
-            <div class="profile__column profile__filed">
-                <h3 class="profile__title profile__title_blue">
-                    {{ title2 }}
-                    <span class="icon icon__add  icon__size_large-x icon_color_red"></span>
-                </h3>
-                <div class="profile__photos">
-                    <img class="profile__photo_img" src="{{ img }}">
-                </div>
-                {{#if edit}}
-                    <input id="photoUpload" type="file">
-                {{/if}}
-            </div>
-            <div class="profile__column profile__filed">
-                <h3 class="profile__title profile__title_blue">
-                    {{ title3 }}
-                    <span class="icon icon__add  icon__size_large-x icon_color_red"></span>
-                </h3>
-                <div class="profile__tags">
-                    {{#if tags}}
-                        {{#each tags as | tag |}}
-                            <div class="tag tag_size_middle">{{ name }}</div>
-                        {{/each}}
-                    {{else}}
-                        <span class="font font__title font__color_lg">У вас пока нет ни одного тэга</span>
-                    {{/if}}
-                </div>
-                <h3 class="profile__title profile__title_blue">
-                    {{ title4 }}
-                    <span class="icon icon__add  icon__size_large-x icon_color_red"></span>
-                </h3>
-                <div class="profile__events">
-                    {{#if events}}
-                        {{#each events as |event|}}
-                            {{>event}}
-                        {{/each}}
-                    {{else}}
-                        <span class="font font__title font__color_lg">Вы пока никуда не ходили</span>
-                    {{/if}}
-                </div>
-            </div>
-        </div>
-    </div>
-</div>
-
