+++ public/js/controllers/profile-controller.js
@@ -1,14 +1,18 @@
 'use strict';
 
-import Controller from '../core/controller.js';
 import ProfileView from '../views/profile-view.js';
-import ProfileEditView from '../views/profile-edit-view.js';
+import MyController from './my-controller.js';
 import UserModel from '../models/user-model.js';
+import ProfileEditView from '../views/profile-edit-view.js';
+import ModalView from '../views/modal-view.js';
+import {tags} from "../utils/static-data.js";
+import {highlightTag} from '../utils/tag-logic.js';
+import logoutRedirect from '../utils/logout.js';
 
 /**
  * @class ProfileController
  */
-export default class ProfileController extends Controller {
+export default class ProfileController extends MyController {
 
     /**
      * construct object of ProfileController class
@@ -20,14 +24,23 @@ export default class ProfileController extends Controller {
         this.editView = null;
         this.image = '';
         this.user = null;
+        this.activeModalWindow = null;
+        this.localTags = [...tags];
+      
+        // TODO: kill meeee
+        document.addEventListener('DOMContentLoaded', () => {
+            this._highlightCircle(2);
+        });
     }
 
+    /**
+     * Create action
+     */
     action() {
         super.action();
         // todo: check is user allowed to see this
         UserModel.getProfile()
             .then((profile) => {
-                console.log(profile);
                 if (!profile) {
                     console.error('Server error');
                     console.log(profile);
@@ -36,61 +49,103 @@ export default class ProfileController extends Controller {
                 if (Object.prototype.hasOwnProperty.call(profile, 'about')) {
                     this.view.render(profile);
                     this.user = profile;
+
                     const photoInput = document.getElementById('photoUpload');
                     photoInput.addEventListener('change', this.#handleFile.bind(this), false);
-                    const textInput = document.getElementsByClassName('btn btn_color_ok btn_size_middle')[0];
-                    console.log(textInput);
+                    const textInput = document.getElementsByClassName('re_btn re_btn__filled')[0];
                     textInput.addEventListener('click', this.#handleInfo.bind(this), false);
-                    const settings = document.getElementsByClassName('btn btn_color_b btn_size_middle')[0];
+                    document.querySelector('.tags_redirect').addEventListener(
+                        'click', this.#showModalTags.bind(this), false);
+                    // TODO: i dunno how to get last item to remove kek in the future
+                    const settings = document.getElementsByClassName('re_btn re_btn__outline kek')[0];
                     settings.addEventListener('click', this.#profileSettings.bind(this), false);
-
+                    document.querySelector('.profile-left__tags').addEventListener(
+                        'click', this.#removeTag, false);
+                    document.getElementsByClassName('re_btn re_btn__outline logout')[0].addEventListener(
+                        'click', logoutRedirect, false);
                 } else {
-                    this.view.render({
-                        name: 'this.user.name',
-                        phone: 'this.user.phone',
-                        email: 'this.user.email',
-                        password: '',
-                        avatar: '/ProfilePhotos/1.jpg',
-                        photos: ['/ProfilePhotos/1.jpg', '/ProfilePhotos/2.jpg'],
-                        gender: 'this.user.gender',
-                        about: 'textInput.value',
-                        rating: 228.1488,
-                        location: {lat: 228.1488, lng: 228.1488, accuracy: 228},
-                        birthday: '2020-02-28T13:55:04.306347+03:00',
-                    });
                     console.error('You have no rights');
                     console.log(profile);
                 }
             }).catch(onerror => {
             console.error(onerror);
         });
-        this.view.render();
     }
 
     #handleFile = (event) => {
         if (event.target.files && event.target.files[0]) {
             let FR = new FileReader();
-            FR.addEventListener('load', this.#photoUploadHandler.bind(this));
+            FR.addEventListener('load', this.#handleSelectImg.bind(this));
             FR.readAsDataURL(event.target.files[0]);
         }
     };
 
+    /**
+     * Preview selected image and draw manage buttons
+     * @param {Event} event
+     */
+    #handleSelectImg = (event) => {
+        event.preventDefault();
+        console.log(event.target);
+        this.image = event.target.result;
+        const photoColumn = document.getElementsByClassName('photo-columns')[0];
+        const newImage = document.createElement('IMG');
+        newImage.src = event.target.result;
+        newImage.className = 'photo';
+        photoColumn.insertAdjacentElement('afterbegin', newImage);
+        const submit = this.#drawButtons('Подтвердить', '2px', true);
+        const discard = this.#drawButtons('Отменить', '2px');
+
+        const text = document.getElementsByClassName('font font_bold font__size_middle font__color_lg')[0];
+        if (text !== undefined) {
+            text.hidden = true;
+        }
+
+        newImage.insertAdjacentElement('afterend', discard);
+        newImage.insertAdjacentElement('afterend', submit);
+        discard.addEventListener('click', (event) => {
+            event.preventDefault();
+            newImage.remove();
+            submit.remove();
+            submit.removeEventListener('click', this.#photoUploadHandler);
+            discard.remove();
+        });
+
+        submit.addEventListener('click', this.#photoUploadHandler.bind(this)); // this bind is really necessary
+    };
+
+    /**
+     * Get image buttons objects
+     * @param {string} title - buttons title
+     * @param {string} margin - margin
+     * @param {boolean} first - is this button first and need 18px margin-left
+     * @returns {HTMLElement} - generated button
+     */
+    #drawButtons = (title, margin, first = false) => {
+        const button = document.createElement('BUTTON');
+        button.className = 're_btn re_btn__outline drawButtonIdentifier';
+        button.innerText = title;
+        button.style.margin = first ? `0 ${margin} 0 18px` : margin;
+        return button;
+    };
+
     #handleInfo = (event) => {
         event.preventDefault();
-        const textInput = document.getElementsByClassName('input input__text_small')[0];
+
+        // Set options
+        const textInput = document.getElementsByClassName('feed__options_field_textarea')[0];
+        const tags = document.querySelectorAll('.tag');
+        let selectedTags = [];
+        tags.forEach((tag) => {
+            selectedTags.push(tag.id);
+        });
         const userProfile = {
-            name: this.user.name,
-            phone: this.user.phone,
-            email: this.user.email,
-            password: '',
-            avatar: {path: this.user.avatar.path},
-            photos: [{path: this.user.avatar.path}],
-            gender: this.user.gender,
-            about: textInput.value,
-            rating: 228.1488,
-            location: {lat: 228.1488, lng: 228.1488, accuracy: 228},
-            birthday: '2020-02-28T13:55:04.306347+03:00',
+            tags: selectedTags,
+            about: textInput.value, // TODO: check if it's safe
+            social: this.user.links,
         };
+
+        // Send request
         UserModel.putProfile(userProfile)
             .then(response => {
                 console.log('ok', response);
@@ -98,33 +153,140 @@ export default class ProfileController extends Controller {
             .catch(reason => console.log('ERROR', reason));
     };
 
+    /**
+     * Upload event photo to server
+     * @param {Event} event
+     */
     #photoUploadHandler = (event) => {
-        this.image = event.target.result;
-        console.log(this.image);
         const userPhoto = this.image.split(';')[1].split(',')[1];
         const userProfile = {
-            name: this.user.name,
-            phone: this.user.phone,
-            email: this.user.email,
-            password: '',
-            avatar: {img: userPhoto},
-            photos: [{img: userPhoto}],
-            gender: this.user.gender,
-            about: this.user.about,
-            rating: 228.1488,
-            location: {lat: 228.1488, lng: 228.1488, accuracy: 228},
-            birthday: '2020-02-28T13:55:04.306347+03:00',
+            uid: this.user.uid,
+            uploaded: {img: userPhoto},
         };
-        UserModel.putProfile(userProfile)
-            .then(response =>
-                document.getElementsByClassName('profile__photo_img')[0].src = this.image)
-            .catch(reason => console.log('ERROR'));
+        document.getElementsByClassName('drawButtonIdentifier')[1].remove();
+        document.getElementsByClassName('drawButtonIdentifier')[0].remove();
+        UserModel.putImage(userProfile)
+            .then(response => {
+                document.getElementsByClassName('photo')[0].src = this.image;
+            }).catch(reason => console.log('ERROR'));
     };
 
+    /**
+     * Show modal window with tags settings
+     * @param event
+     */
+    #showModalTags = (event) => {
+        this.editView = new ModalView(document.body);
+
+        // Rendering active tags in modal view
+        let activeTags = document.body.querySelectorAll('.tag__container.tag__container__active');
+        let activeTagsTitles = [];
+        for (let iii = 0; iii < activeTags.length; iii++) {
+            activeTagsTitles.push(activeTags[iii].firstElementChild.innerText);
+        }
+        this.localTags.forEach((tag) => {
+            tag.editable = true;
+            if (activeTagsTitles.includes(tag.title)) {
+                tag.active_class = 'tag__container__active';
+            }
+        });
+        this.editView.render({
+            title: 'Ваши теги',
+            tags: this.localTags,
+            last_buttons: [
+                {title: 'Сохранить',}]
+            });
+        let modalBG = document.body.querySelector('.modal__bg');
+        modalBG.addEventListener('click', (event) => {
+            event.preventDefault();
+            event.stopPropagation();
+        });
+        this.activeModalWindow = modalBG.firstElementChild;
+        this.activeModalWindow.querySelector('.modal__body').addEventListener(
+            'click', highlightTag, false);
+        this.activeModalWindow.querySelector('.modal__header__icon').addEventListener(
+            'click', (event) => {
+                event.preventDefault();
+                this.editView.clear();
+                this.editView = null;});
+        this.activeModalWindow.querySelector(
+            '.modal__footer').querySelector(
+                '.re_btn.re_btn__outline').addEventListener(
+                    'click', this.#submitTagsHandler.bind(this), false)
+    };
+
+    #submitTagsHandler = (event) => {
+        event.preventDefault();
+
+        let tagsField = document.body.querySelector('.profile-left__tags');
+        let prevLength = tagsField.length;
+
+        // Remove all tagsField children
+        while (tagsField.lastElementChild) {
+            tagsField.removeChild(tagsField.lastElementChild);
+        }
+
+        let allTags = this.activeModalWindow.querySelectorAll('.tag__container');
+        this.localTags = [];
+        let length = 0;
+        allTags.forEach((tag) => {
+            let tempTag = {
+                title: tag.firstElementChild.innerText,
+                editable: true,
+            };
+            if (tag.classList.contains('tag__container__active')) {
+                tempTag.active_class = 'tag__container__active';
+                tagsField.appendChild(tag);
+                length++;
+            }
+            this.localTags.push(tempTag);}
+        );
+
+        if (length === 0) {
+            let emptyMessageText = (prevLength !== 0)
+                ? 'Вы удалили все теги'
+                : 'У вас пока нет ни одного тэга';
+            // TODO: replace with HBS block
+            let emptyMessage = document.createElement('div');
+            emptyMessage.classList.add('center');
+            let message = document.createElement('span');
+            message.classList.add("font", "font_bold", "font__size_small", "font__color_lg");
+            message.innerText = emptyMessageText;
+            emptyMessage.appendChild(message);
+            tagsField.appendChild(emptyMessage);
+        }
+
+        this.editView.clear();
+        this.editView = null;
+    };
+
+    #removeTag = (event) => {
+        let elem = event.target;
+        let elemContainer = elem.closest('.tag__container');
+        if (elemContainer && elemContainer.classList.contains('tag__container')) {
+            delete this.localTags.find((tag) => {
+                return (tag.title === elemContainer.firstElementChild.innerText);
+            }).active_class;
+
+            // Check if it was the last tag
+            if (elemContainer.parentElement.childElementCount === 1) {
+                // TODO: replace with HBS block
+                let emptyMessage = document.createElement('div');
+                emptyMessage.classList.add('center');
+                let message = document.createElement('span');
+                message.classList.add("font", "font_bold", "font__size_small", "font__color_lg");
+                message.innerText = 'Вы удалили все теги';
+                emptyMessage.appendChild(message);
+                elemContainer.parentElement.appendChild(emptyMessage);
+            }
+          
+            elemContainer.remove();
+        }
+    };
 
     /**
      * Create profile settings popup
-     * @param event
+     * @param {Event} event
      */
     #profileSettings = (event) => {
         this.editView = new ProfileEditView(this.parent);
@@ -138,7 +300,7 @@ export default class ProfileController extends Controller {
 
     /**
      * Remove profile settings popup
-     * @param event
+     * @param {Event} event
      */
     #removeProfileSettings = (event) => {
         const popup = document.getElementsByClassName('profile-edit')[0];
