+++ public/js/index.js
@@ -1,12 +1,11 @@
 'use strict';
 
 import LandingController from './controllers/landing-controller.js';
-import ProfileController from './controllers/profile-controller.js';
 import LoginController from './controllers/login-controller.js';
 import SignUpController from './controllers/signup-controller.js';
 import FeedUsersController from './controllers/feed-users-controller.js';
 import BigEventSearchController from './controllers/big-event-search-controller.js';
-import NewProfileController from './controllers/new-profile-controller.js';
+import ProfileController from './controllers/profile-controller.js';
 import Router from './core/router.js';
 import '../static/css/style.css';
 
@@ -15,10 +14,10 @@ let application = document.getElementById('application');
 let router = new Router();
 router.addRoute('/', new LandingController(application));
 router.addRoute('/login', new LoginController(application));
-router.addRoute('/profile', new ProfileController(application));
 router.addRoute('/signup', new SignUpController(application));
-router.addRoute('/search', new BigEventSearchController(application));         // big events
-router.addRoute('/feed/users', new FeedUsersController(application));          // profiles
-router.addRoute('/my/profile', new NewProfileController(application));
+router.addRoute('/search', new BigEventSearchController(application));  // big & middle events
+router.addRoute('/feed/users', new FeedUsersController(application));   // profiles
+router.addRoute('/feed/events', new FeedUsersController(application));  // events
+router.addRoute('/my/profile', new ProfileController(application));
 
 router.route();
